refactor: split sharded pre-pass into dedicated modules (module-size ratchet)
The CI Rust gate failed on the size ratchet: parallel_scan.rs crossed 400
lines and both prepass.rs files blew their budgets. Split along the seams
the sharded pre-pass introduced (AGENTS.md rule — split, don't raise):

- rust/processing/src/shard_classes.rs: PREPASS_CLASS_* codes +
  scan_shard_classified (parallel_scan.rs keeps the byte-identical
  shard/stitch protocol).
- rust/processing/src/prepass_styled.rs: StyleSeeds +
  resolve_styled_items_into (re-exported from prepass for callers).
- rust/wasm-bindings/.../prepass_sharded.rs: scanEntityIndexShard,
  resolveStyledItemsShard, finalizePrepassStyles,
  buildPrePassStreamingSharded, plus the SHARED wire serializers
  (set_stream_meta_props, styles_payload — the styles payload stanza was
  previously duplicated between the serial emission and finalize; now one
  fn, so the two cannot drift).

Behavior identical: sharded smoke reproduces exact parity
(1960 styled / 83 draws / 12.7MB on FM_ARC_DigitalHub); ratchet, clippy
-D warnings, and the 51 processing lib tests all green locally.
