+++ packages/wasm/pkg/ifc-lite.d.ts
@@ -156,6 +156,61 @@ export class IfcAPI {
    */
   exportObj(content: Uint8Array, include_normals: boolean, hidden: Uint32Array, isolated: Uint32Array): Uint8Array;
   /**
+   * Sharded pre-pass: merge the shard-resolved styled-item columns with the
+   * SUPPORT spans (extracted host-side from the shard classes) and run the
+   * CANONICAL styles flatten. Returns the exact `styles` event payload the
+   * serial path emits. Runs on any worker with `setEntityIndex` installed.
+   * Span arguments are `[id, start, len]` triples; `plane_angle_to_radians`
+   * comes from the meta event.
+   */
+  finalizePrepassStyles(data: Uint8Array, orphan_ids: Uint32Array, orphan_colors: Float32Array, geom_ids: Uint32Array, geom_colors: Float32Array, colour_map_spans: Uint32Array, material_def_spans: Uint32Array, rel_material_spans: Uint32Array, void_spans: Uint32Array, fills_spans: Uint32Array, aggregate_spans: Uint32Array, plane_angle_to_radians: number): any;
+  /**
+   * SPIKE (sharded pre-pass): scan the entity index over a single byte range.
+   *
+   * Each idle browser geometry worker calls this on its `[range_start,
+   * range_end)` shard; the main thread stitches the returned columns into the
+   * full entity index (byte-identical to the single-threaded
+   * `build_entity_index`) by binary-searching each shard for the previous
+   * shard's `handoff`. Delegates to `ifc_lite_processing::scan_shard`, the
+   * exact per-chunk primitive the native `build_entity_index_parallel` fans
+   * across cores — so the sharded merge cannot drift from the serial builder.
+   *
+   * Byte offsets returned are GLOBAL (relative to file start), so shards
+   * concatenate without rewriting. Returns a plain object:
+   *   `{ ids: Uint32Array, starts: Uint32Array, lengths: Uint32Array,
+   *      handoff: number }`
+   * where `handoff` is the global start of the first entity at/after
+   * `range_end` (the next shard's first real entity), or `-1` at EOF.
+   */
+  scanEntityIndexShard(data: Uint8Array, range_start: number, range_end: number): any;
+  /**
+   * Sharded pre-pass: resolve ONE contiguous (file-ordered) slice of the
+   * styled-item span list on this worker, against the entity index installed
+   * by `setEntityIndex`. Returns raw resolved maps as flat columns:
+   * `{ orphanIds, orphanColors (f32 rgba per id), geomIds, geomColors }`.
+   * The host merges shard results IN SHARD ORDER with first-wins per
+   * geometry id, reproducing the serial resolver's file-order precedence,
+   * then hands the merged columns to `finalizePrepassStyles`.
+   * `spans` is `[id, start, len]` triples.
+   */
+  resolveStyledItemsShard(data: Uint8Array, spans: Uint32Array): any;
+  /**
+   * Sharded pre-pass variant: same scan/discovery/jobs/columns pipeline as
+   * `buildPrePassStreaming`, but
+   *  1. the entity index is PREBUILT from the host's stitched shard columns
+   *     (file order; see `scanEntityIndexShard`) — the scan skips its inline
+   *     index build, the meta RTC ladder resolves against the FULL index
+   *     (no partial-ladder full-rescan detour), and the post-scan
+   *     `entity-index` event is skipped (the host already delivered it), and
+   *  2. styles resolution is EXTERNAL: the styled-item spans are resolved as
+   *     shard slices on the geometry workers (`resolveStyledItemsShard`);
+   *     this call stashes the SUPPORT spans + plane-angle scale, and the
+   *     follow-up `finalizePrepassStyles` merges + flattens into the exact
+   *     styles payload the serial path emits. NO `styles` event is emitted
+   *     here.
+   */
+  buildPrePassStreamingSharded(data: Uint8Array, on_event: Function, chunk_size: number, disabled_type_names: string[] | null | undefined, skip_type_geometry: boolean, index_ids: Uint32Array, index_starts: Uint32Array, index_lengths: Uint32Array): any;
+  /**
    * Store the whole IFC source file ONCE per load so the `*FromSource` batch
    * variants can read it from the wasm heap instead of re-copying it per call.
    *
@@ -227,34 +282,6 @@ export class IfcAPI {
    */
   buildPrePassOnce(data: Uint8Array): any;
   /**
-   * Sharded pre-pass: merge the shard-resolved styled-item columns with the
-   * SUPPORT spans (extracted host-side from the shard classes) and run the
-   * CANONICAL styles flatten. Returns the exact `styles` event payload the
-   * serial path emits. Runs on any worker with `setEntityIndex` installed.
-   * Span arguments are `[id, start, len]` triples; `plane_angle_to_radians`
-   * comes from the meta event.
-   */
-  finalizePrepassStyles(data: Uint8Array, orphan_ids: Uint32Array, orphan_colors: Float32Array, geom_ids: Uint32Array, geom_colors: Float32Array, colour_map_spans: Uint32Array, material_def_spans: Uint32Array, rel_material_spans: Uint32Array, void_spans: Uint32Array, fills_spans: Uint32Array, aggregate_spans: Uint32Array, plane_angle_to_radians: number): any;
-  /**
-   * SPIKE (sharded pre-pass): scan the entity index over a single byte range.
-   *
-   * Each idle browser geometry worker calls this on its `[range_start,
-   * range_end)` shard; the main thread stitches the returned columns into the
-   * full entity index (byte-identical to the single-threaded
-   * `build_entity_index`) by binary-searching each shard for the previous
-   * shard's `handoff`. Delegates to `ifc_lite_processing::scan_shard`, the
-   * exact per-chunk primitive the native `build_entity_index_parallel` fans
-   * across cores — so the sharded merge cannot drift from the serial builder.
-   *
-   * Byte offsets returned are GLOBAL (relative to file start), so shards
-   * concatenate without rewriting. Returns a plain object:
-   *   `{ ids: Uint32Array, starts: Uint32Array, lengths: Uint32Array,
-   *      handoff: number }`
-   * where `handoff` is the global start of the first entity at/after
-   * `range_end` (the next shard's first real entity), or `-1` at EOF.
-   */
-  scanEntityIndexShard(data: Uint8Array, range_start: number, range_end: number): any;
-  /**
    * Streaming pre-pass: emits geometry jobs in chunks via a JS callback
    * instead of waiting for the full file scan to complete.
    *
@@ -280,33 +307,6 @@ export class IfcAPI {
    *   `{ type: "complete", totalJobs }`
    */
   buildPrePassStreaming(data: Uint8Array, on_event: Function, chunk_size: number, disabled_type_names: string[] | null | undefined, skip_type_geometry: boolean): any;
-  /**
-   * Sharded pre-pass: resolve ONE contiguous (file-ordered) slice of the
-   * styled-item span list on this worker, against the entity index installed
-   * by `setEntityIndex`. Returns raw resolved maps as flat columns:
-   * `{ orphanIds, orphanColors (f32 rgba per id), geomIds, geomColors }`.
-   * The host merges shard results IN SHARD ORDER with first-wins per
-   * geometry id, reproducing the serial resolver's file-order precedence,
-   * then hands the merged columns to `finalizePrepassStyles`.
-   * `spans` is `[id, start, len]` triples.
-   */
-  resolveStyledItemsShard(data: Uint8Array, spans: Uint32Array): any;
-  /**
-   * Sharded pre-pass variant: same scan/discovery/jobs/columns pipeline as
-   * `buildPrePassStreaming`, but
-   *  1. the entity index is PREBUILT from the host's stitched shard columns
-   *     (file order; see `scanEntityIndexShard`) — the scan skips its inline
-   *     index build, the meta RTC ladder resolves against the FULL index
-   *     (no partial-ladder full-rescan detour), and the post-scan
-   *     `entity-index` event is skipped (the host already delivered it), and
-   *  2. styles resolution is EXTERNAL: the styled-item spans are resolved as
-   *     shard slices on the geometry workers (`resolveStyledItemsShard`);
-   *     this call stashes the SUPPORT spans + plane-angle scale, and the
-   *     follow-up `finalizePrepassStyles` merges + flattens into the exact
-   *     styles payload the serial path emits. NO `styles` event is emitted
-   *     here.
-   */
-  buildPrePassStreamingSharded(data: Uint8Array, on_event: Function, chunk_size: number, disabled_type_names: string[] | null | undefined, skip_type_geometry: boolean, index_ids: Uint32Array, index_starts: Uint32Array, index_lengths: Uint32Array): any;
   /**
    * Parse the file and return structured per-axis data (tag + endpoints) in
    * the renderer's Y-up world space (RTC-subtracted, metres). Use this when
